+++ hawkbit-ddi/hawkbit-ddi-starter/pom.xml
@@ -58,11 +58,6 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.hawkbit</groupId>
-            <artifactId>hawkbit-http-security</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.hawkbit</groupId>
             <artifactId>hawkbit-repository-jpa</artifactId>
             <version>${project.version}</version>
         </dependency>
