+++ hawkbit-ddi/hawkbit-ddi-starter/src/main/java/org/eclipse/hawkbit/autoconfigure/ddi/DdiApiAutoConfiguration.java
@@ -20,6 +20,6 @@ import org.springframework.context.annotation.Import;
 @Configuration
 @ConditionalOnClass(DdiApiConfiguration.class)
 @Import(DdiApiConfiguration.class)
-public class DDiApiAutoConfiguration {
+public class DdiApiAutoConfiguration {
 
 }
