+++ hawkbit-ddi/hawkbit-ddi-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,4 @@
-org.eclipse.hawkbit.autoconfigure.ddi.DDiApiAutoConfiguration
+org.eclipse.hawkbit.autoconfigure.ddi.DdiApiAutoConfiguration
+org.eclipse.hawkbit.rest.SecurityManagedConfiguration
+org.eclipse.hawkbit.autoconfigure.ddi.ControllerSecurityConfiguration
+org.eclipse.hawkbit.autoconfigure.ddi.ControllerDownloadSecurityConfiguration
