+++ hawkbit-mgmt/hawkbit-mgmt-starter/pom.xml
@@ -35,11 +35,6 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.hawkbit</groupId>
-            <artifactId>hawkbit-http-security</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.hawkbit</groupId>
             <artifactId>hawkbit-repository-jpa</artifactId>
             <version>${project.version}</version>
         </dependency>
