+++ hawkbit-mgmt/hawkbit-mgmt-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,3 @@
 org.eclipse.hawkbit.autoconfigure.mgmt.MgmtApiAutoConfiguration
+org.eclipse.hawkbit.rest.SecurityManagedConfiguration
+org.eclipse.hawkbit.autoconfigure.mgmt.MgmtSecurityConfiguration
