+++ hawkbit-test-report/pom.xml
@@ -38,11 +38,6 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.hawkbit</groupId>
-            <artifactId>hawkbit-http-security</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.hawkbit</groupId>
             <artifactId>hawkbit-repository-api</artifactId>
             <version>${project.version}</version>
         </dependency>
