+++ site/src/main/java/org/eclipse/hawkbit/doc/Start.java
@@ -9,7 +9,6 @@
  */
 package org.eclipse.hawkbit.doc;
 
-import org.eclipse.hawkbit.autoconfigure.security.EnableHawkbitManagedSecurityConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -18,7 +17,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * The minimal configuration for the stand alone hawkBit server.
  */
 @SpringBootApplication
-@EnableHawkbitManagedSecurityConfiguration
 // Exception squid:S1118 - Spring boot standard behavior
 @SuppressWarnings({ "squid:S1118" })
 public class Start {
